requirements.txt
opencv-python>=4.5.0
pillow>=8.0.0
pyyaml>=5.4.0 